+++ addon.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <addon
     id="plugin.video.picta"
-    version="1.0.4"
+    version="1.1.0"
     name="Picta"
     provider-name="crmartinez, oleksis">
     <requires>
@@ -21,14 +21,14 @@
         <license>MIT</license>
         <website>https://ww.picta.cu</website>
         <source>https://github.com/oleksis/plugin.video.picta/tree/kodi_19</source>
-        <news>1.0.4 (2021-06-17)
-Add requires InputStream: Adaptative, FFmpeg Direct
-Add screenshots
+        <news>1.1.0 (2021-06-18)
+Restructure addon.
         </news>
         <assets>
             <icon>resources/icon.png</icon>
             <fanart>resources/fanart.jpg</fanart>
             <screenshot>resources/images/screenshot-01.png</screenshot>
+            <screenshot>resources/images/screenshot-02.png</screenshot>
         </assets>
     </extension>
 </addon>
